Thingvisors/DockerThingVisor/ThingVisor_CBPF/v_camera_system/src/requirements.txt
setuptools~=52.0.0
opencv-python~=4.5.1.48
paho-mqtt~=1.5.1
numpy~=1.19.2
msrest~=0.6.21
Pillow~=8.1.0
Werkzeug~=1.0.1
Flask~=1.1.2
azure-cognitiveservices-vision-face
